Rename examples.
