+++ examples/reason-react-example/src/interop/InteropRoot.js
@@ -17,7 +17,7 @@ const helloWorld = GreetingRe.concat("++", helloWorldList);
 
 const someNumber: number = GreetingRe.testDefaultArgs({ y: 10 });
 
-const TestRE = require("./Test.re");
+const TestRE = require("./WrapJsValue.re");
 console.log("interopRoot.js roundedNumber:", TestRE.roundedNumber);
 
 const App = () => (
