+++ examples/reason-react-example/src/interop/WrapJsValue.bs.js
@@ -1,9 +1,9 @@
 // Generated by BUCKLESCRIPT VERSION 4.0.6, PLEASE EDIT WITH CARE
 'use strict';
 
-var TestRe = require("./Test.re");
+var WrapJsValueRe = require("./WrapJsValue.re");
 
-var roundedNumber = TestRe.round(1.8);
+var roundedNumber = WrapJsValueRe.round(1.8);
 
 console.log("Test.re roundedNumber:", roundedNumber);
