+++ examples/reason-react-example/src/interop/WrapJsValue.re.js
@@ -10,6 +10,6 @@ const MyMath = require("./MyMath");
 export const round: (number) => number = MyMath.round;
 
 // $FlowExpectedError: Reason checked type sufficiently
-const TestBS = require("./Test.bs");
+const WrapJsValueBS = require("./WrapJsValue.bs");
 
-export const roundedNumber: number = TestBS.roundedNumber;
+export const roundedNumber: number = WrapJsValueBS.roundedNumber;
