+++ examples/typescript-react-example/src/WrapJsValue.bs.js
@@ -1,9 +1,9 @@
 // Generated by BUCKLESCRIPT VERSION 4.0.6, PLEASE EDIT WITH CARE
 'use strict';
 
-var Test = require("./Test");
+var WrapJsValue = require("./WrapJsValue");
 
-var roundedNumber = Test.round(1.8);
+var roundedNumber = WrapJsValue.round(1.8);
 
 console.log("Test.re roundedNumber:", roundedNumber);
