+++ examples/typescript-react-example/src/index.tsx
@@ -31,7 +31,7 @@ const businesses = [
 
 const addresses = Records.findAllAddresses(businesses);
 
-import {roundedNumber} from "./Test";
+import {roundedNumber} from "./WrapJsValue";
 // tslint:disable-next-line:no-console
 console.log("interopRoot.js roundedNumber:", roundedNumber);
